v2026010 – ESPHome 2026.1 Compatibility and Climate Page Fixes

+++ components/nspanel_ha_blueprint/page_climate.h
@@ -15,7 +15,7 @@ namespace nspanel_ha_blueprint {
     extern float set_climate_target_temp_high;
     extern float set_climate_target_temp_low;
     extern uint8_t set_climate_temp_step;
-    extern uint8_t set_climate_total_steps;
+    extern uint16_t set_climate_total_steps;
     extern uint8_t set_climate_temp_offset;
     extern std::string set_climate_climate_icon;
     extern bool set_climate_embedded_climate;
